components/form-expert/requirements.txt
bottle==0.12.16
python-Levenshtein==0.12.0
nltk==3.4
spacy==2.0.18
pymongo==3.7.2
pytest==4.2.0